restructured package to allow for run directly from command line
